Extract shared OpenAI transport from the LLM client
Image encoding, user-payload building, response text decode, and lazy SDK
loading move to models/llm/openai_transport.py so the remote LibreVLM
transport can reuse them. No behavior change; client re-exports the names
its tests patch.
